Make the notification list only cover 90% height
If you open the notification list on small pages it can get higher than
the browser window if you have many notification. This would cause
double scollbars when only one is necessary. By changing the maximum
height to 90% there appears a small grey bar at the bottom of the
notification bar, but it should fix the problem with new extra
scrollbars appearing for most screen sizes.