Fix #737, place responsive css before bootstrap
The default config for concatenating the vendor css files does not seem
to specify an order. However, for the responsive features of bootstrap
to take effect, the file `bootstrap-responsive.css` must follow the file
`bootstrap.css` in the source. This commit makes the files and their
order explicit.

The default vendor config can be seen here:
https://github.com/testdouble/lineman/blob/master/config/files.coffee