Don't treat instance variables (in code blocks) as mentions
For lack of the knowledge of a simpler way and because I did't want to
write code to parse markdown code blocks myself, I implemented this by
rendering the markdown to HTML using the same Markdown processor
that the site uses and then using Nokogiri to strip out any <code>
or <td class="code"> nodes, which were the two ways that the code spans
and fenced code blocks were being rendered, respectively.

A couple of potential downsides to this method are: 1. If the rendering
of code blocks changes this will no longer work. (However the tests
should catch it). and 2. I'm not sure how fast the markdown renderer is
and how fast Nokogiri is at doing these operations, so I'm not sure if
this could be a potential performance hit. I suspect it won't be
significant though.